requirements.txt
numpy>=1.17.0
pandas>=1.0,!=1.1 #  required by Anndata
anndata>=0.7.4
# h5py<3.0.0 #  avoid byte strings but caused building errors
# h5py>=3.4
scikit-learn>=0.19
scipy>=1.4
kneed>=0.7
seaborn>=0.11
matplotlib>=3.3
adjusttext>=0.7.3
umap-learn>=0.3.0
tables
tqdm>=4.0
torchbiggraph
